+++ ado_wrapper/__init__.py
@@ -7,7 +7,7 @@ __all__ = [
     "AgentPool", "AnnotatedTag", "Artifact", "AuditLog", "Branch", "BuildTimeline", "Build", "BuildDefinition", "Commit",
     "Environment", "PipelineAuthorisation", "Group", "HierarchyCreatedBuildDefinition", "MergeBranchPolicy", "MergePolicies",
     "MergePolicyDefaultReviewer", "MergeTypeRestrictionPolicy", "Organisation", "Permission", "PersonalAccessToken", "Project",
-    "PullRequest", "PullRequestComment", "Release", "ReleaseDefinition", "RepoUserPermissions", "UserPermission",
+    "PullRequest", "PullRequestCommentThread", "PullRequestComment", "Release", "ReleaseDefinition", "RepoUserPermissions", "UserPermission",
     "BuildRepository", "Repo", "Run", "CodeSearch", "SecureFile", "ServiceEndpoint", "Team", "AdoUser", "Member", "Reviewer", "TeamMember",
     "VariableGroup",
 ]  # fmt: skip