+++ ado_wrapper/resources/annotated_tags.py
@@ -67,6 +67,7 @@ class AnnotatedTag(StateManagedResource):
 
     def link(self, ado_client: "AdoClient") -> str:
         return f"https://dev.azure.com/{ado_client.ado_org_name}/{ado_client.ado_project_name}/_git/{self.repo_id}?version=GT{self.name}"
+
     # # ============ End of requirement set by all state managed resources ================== #
     # # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #
     # # =============== Start of additional methods included with class ===================== #