+++ ado_wrapper/resources/groups.py
@@ -55,6 +55,7 @@ class Group(StateManagedResource):
 
     def link(self, ado_client: "AdoClient") -> str:
         return f"https://dev.azure.com/{ado_client.ado_org_name}/{ado_client.ado_project_name}/_settings/permissions?subjectDescriptor={self.group_descriptor}"
+
     # ============ End of requirement set by all state managed resources ================== #
     # ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ #
     # =============== Start of additional methods included with class ===================== #